requirements.txt
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.5 
gunicorn==20.1.0
django-heroku==0.3.1
django-cors-headers==3.13.0